Shared generic pipeline components (config, tokenziers) moved to common module
