+++ src/pipelines/token_classification.rs
@@ -15,12 +15,13 @@
 //! More generic token classification pipeline, works with multiple models (Bert, Roberta)
 //!
 //! ```no_run
-//! use rust_bert::pipelines::token_classification::{ModelType,TokenClassificationModel,TokenClassificationConfig};
+//! use rust_bert::pipelines::token_classification::{TokenClassificationModel,TokenClassificationConfig};
 //! use rust_bert::resources::{Resource,RemoteResource};
+//! use rust_bert::bert::{BertModelResources, BertVocabResources, BertConfigResources};
+//! use rust_bert::pipelines::common::ModelType;
 //!# fn main() -> failure::Fallible<()> {
 //!
 //! //Load a configuration
-//! use rust_bert::bert::{BertModelResources, BertVocabResources, BertConfigResources};
 //! let config = TokenClassificationConfig::new(ModelType::Bert,
 //!    Resource::Remote(RemoteResource::from_pretrained(BertModelResources::BERT_NER)),
 //!    Resource::Remote(RemoteResource::from_pretrained(BertVocabResources::BERT_NER)),
@@ -45,31 +46,28 @@
 //!# use rust_bert::pipelines::token_classification::Token;
 //!# let output =
 //! [
-//!    Token { text: String::from("Amy"), score: 0.9986, label: String::from("I-PER"), sentence: 0, index: 0}, //, word_index: Some(3), continuation: false },
-//!    Token { text: String::from("Paris"), score: 0.9985, label: String::from("I-LOC"), sentence: 0, index: 9}, // , word_index: Some(8), continuation: false},
-//!    Token { text: String::from("Paris"), score: 0.9988, label: String::from("I-LOC"), sentence: 1, index: 1}, //, word_index: Some(0), continuation: false},
-//!    Token { text: String::from("France"), score: 0.9993, label: String::from("I-LOC"), sentence: 1, index: 6} //, word_index: Some(5), continuation: false},
+//!    Token { text: String::from("Amy"), score: 0.9986, label: String::from("I-PER"), sentence: 0, index: 0},
+//!    Token { text: String::from("Paris"), score: 0.9985, label: String::from("I-LOC"), sentence: 0, index: 9},
+//!    Token { text: String::from("Paris"), score: 0.9988, label: String::from("I-LOC"), sentence: 1, index: 1},
+//!    Token { text: String::from("France"), score: 0.9993, label: String::from("I-LOC"), sentence: 1, index: 6}
 //! ]
 //!# ;
 //! ```
 
-use rust_tokenizers::bert_tokenizer::{BertTokenizer};
-use rust_tokenizers::{RobertaTokenizer};
 use tch::nn::VarStore;
-use rust_tokenizers::preprocessing::tokenizer::base_tokenizer::{Tokenizer,TokenizedInput,TruncationStrategy};
-use std::path::Path;
+use rust_tokenizers::preprocessing::tokenizer::base_tokenizer::{Tokenizer, TokenizedInput, TruncationStrategy};
 use std::collections::HashMap;
 use tch::{Tensor, no_grad, Device};
 use tch::kind::Kind::Float;
-use crate::bert::{BertForTokenClassification, BertConfig, BertModelResources, BertConfigResources, BertVocabResources};
-use crate::roberta::{RobertaForTokenClassification};
-use crate::distilbert::{DistilBertForTokenClassification, DistilBertConfig};
-use crate::Config;
+use crate::bert::{BertForTokenClassification, BertModelResources, BertConfigResources, BertVocabResources};
+use crate::roberta::RobertaForTokenClassification;
+use crate::distilbert::DistilBertForTokenClassification;
 use crate::common::resources::{Resource, RemoteResource, download_resource};
-use serde::{Serialize,Deserialize};
+use serde::{Serialize, Deserialize};
+use crate::pipelines::common::{ModelType, ConfigOption, TokenizerOption};
 
 
-#[derive(Debug,Serialize,Deserialize)]
+#[derive(Debug, Serialize, Deserialize)]
 /// # Token generated by a `TokenClassificationModel`
 pub struct Token {
     /// String representation of the Token
@@ -120,14 +118,14 @@ impl TokenClassificationConfig {
     /// * vocab - An optional `Resource` tuple (`Option<Resource>`) pointing to the tokenizer's merge file to load (e.g.  merges.txt), needed only for Roberta.
     /// * lower_case - A `bool' indicating whether the tokenizer should lower case all input (in case of a lower-cased model)
     ///
-    pub fn new(model_type: ModelType, model: Resource, config: Resource, vocab: Resource, merges: Option<Resource>, lower_case: bool) -> TokenClassificationConfig {
+    pub fn new(model_type: ModelType, model_resource: Resource, config_resource: Resource, vocab_resource: Resource, merges_resource: Option<Resource>, lower_case: bool) -> TokenClassificationConfig {
         TokenClassificationConfig {
-            model_type: model_type,
-            model_resource: model,
-            config_resource: config,
-            vocab_resource: vocab,
-            merges_resource: merges,
-            lower_case: lower_case,
+            model_type,
+            model_resource,
+            config_resource,
+            vocab_resource,
+            merges_resource,
+            lower_case,
             device: Device::cuda_if_available(),
         }
     }
@@ -148,30 +146,6 @@ impl Default for TokenClassificationConfig {
     }
 }
 
-#[derive(Clone,Copy,Serialize,Deserialize)]
-/// # Identifies the type of model
-pub enum ModelType {
-    Bert,
-    DistilBert,
-    Roberta,
-}
-
-/// # Abstraction that holds a model configuration, can be of any of the supported models
-pub enum ConfigOption {
-    /// Bert configuration
-    Bert(BertConfig),
-    /// DistilBert configuration
-    DistilBert(DistilBertConfig),
-}
-
-/// # Abstraction that holds a particular tokenizer, can be of any of the supported models
-pub enum TokenizerOption {
-    /// Bert Tokenizer
-    Bert(BertTokenizer),
-    /// Roberta Tokenizer
-    Roberta(RobertaTokenizer),
-}
-
 /// # Abstraction that holds one particular token sequence classifier model, for any of the supported models
 pub enum TokenClassificationOption {
     /// Bert for Token Classification
@@ -181,51 +155,6 @@ pub enum TokenClassificationOption {
     /// Roberta for Token Classification
     Roberta(RobertaForTokenClassification),
 }
-
-impl ConfigOption {
-    /// Interface method to load a configuration from file
-    pub fn from_file(model_type: ModelType, path: &Path) -> Self {
-        match model_type {
-            ModelType::Bert | ModelType::Roberta => ConfigOption::Bert(BertConfig::from_file(path)),
-            ModelType::DistilBert => ConfigOption::DistilBert(DistilBertConfig::from_file(path))
-        }
-    }
-
-    pub fn get_label_mapping(self) -> HashMap<i64,String> {
-        match self {
-            Self::Bert(config) => config.id2label.expect("No label dictionary (id2label) provided in configuration file"),
-            Self::DistilBert(config) => config.id2label.expect("No label dictionary (id2label) provided in configuration file"),
-        }
-    }
-}
-
-impl TokenizerOption {
-    /// Interface method to load a tokenizer from file
-    pub fn from_file(model_type: ModelType, vocab_path: &str, merges_path: Option<&str>, lower_case: bool) -> Self {
-        match model_type {
-            ModelType::Bert | ModelType::DistilBert => TokenizerOption::Bert(BertTokenizer::from_file(vocab_path,lower_case)),
-            ModelType::Roberta => TokenizerOption::Roberta(RobertaTokenizer::from_file(vocab_path, merges_path.expect("No merges specified!"), lower_case)),
-        }
-    }
-
-    /// Returns the model type
-    pub fn model_type(&self) -> ModelType {
-        match *self {
-            Self::Bert(_) => ModelType::Bert,
-            Self::Roberta(_) => ModelType::Roberta
-        }
-    }
-
-    /// Interface method
-    pub fn encode_list(&self, text_list: Vec<&str>, max_len: usize, truncation_strategy: &TruncationStrategy, stride: usize) -> Vec<TokenizedInput> {
-        match *self {
-            Self::Bert(ref tokenizer) => tokenizer.encode_list(text_list, max_len, truncation_strategy, stride),
-            Self::Roberta(ref tokenizer) => tokenizer.encode_list(text_list, max_len, truncation_strategy, stride)
-        }
-    }
-}
-
-
 
 impl TokenClassificationOption {
     /// Instantiate a new token sequence classification model of the supplied type.
@@ -245,7 +174,7 @@ impl TokenClassificationOption {
                 } else {
                     panic!("You can only supply a BertConfig for Bert!");
                 }
-            },
+            }
             ModelType::DistilBert => {
                 if let ConfigOption::DistilBert(config) = config {
                     TokenClassificationOption::DistilBert(DistilBertForTokenClassification::new(p, config))
@@ -279,23 +208,22 @@ impl TokenClassificationOption {
                      token_type_ids: Option<Tensor>,
                      position_ids: Option<Tensor>,
                      input_embeds: Option<Tensor>,
-                     train: bool) -> (Tensor, Option<Vec<Tensor>>, Option<Vec<Tensor>>)  {
+                     train: bool) -> (Tensor, Option<Vec<Tensor>>, Option<Vec<Tensor>>) {
         match *self {
-            Self::Bert(ref model) => model.forward_t(input_ids,mask,token_type_ids,position_ids,input_embeds,train),
-            Self::DistilBert(ref model) => model.forward_t(input_ids,mask,input_embeds,train).expect("Error in distilbert forward_t"),
-            Self::Roberta(ref model) => model.forward_t(input_ids,mask,token_type_ids,position_ids,input_embeds,train),
+            Self::Bert(ref model) => model.forward_t(input_ids, mask, token_type_ids, position_ids, input_embeds, train),
+            Self::DistilBert(ref model) => model.forward_t(input_ids, mask, input_embeds, train).expect("Error in distilbert forward_t"),
+            Self::Roberta(ref model) => model.forward_t(input_ids, mask, token_type_ids, position_ids, input_embeds, train),
         }
-
     }
-
 }
 
 
 /// # TokenClassificationModel for Named Entity Recognition or Part-of-Speech tagging
 pub struct TokenClassificationModel {
     tokenizer: TokenizerOption,
-    token_sequence_classifier: TokenClassificationOption, //e.g. BertForTokenClassification,
-    label_mapping: HashMap<i64,String>,
+    token_sequence_classifier: TokenClassificationOption,
+    //e.g. BertForTokenClassification,
+    label_mapping: HashMap<i64, String>,
     var_store: VarStore,
 }
 
@@ -339,9 +267,9 @@ impl TokenClassificationModel {
 
     fn prepare_for_model(&self, input: Vec<&str>) -> Tensor {
         let tokenized_input: Vec<TokenizedInput> = self.tokenizer.encode_list(input.to_vec(),
-                                                         128,
-                                                         &TruncationStrategy::LongestFirst,
-                                                         0);
+                                                                              128,
+                                                                              &TruncationStrategy::LongestFirst,
+                                                                              0);
         let max_len = tokenized_input.iter().map(|input| input.token_ids.len()).max().unwrap();
         let tokenized_input_tensors: Vec<tch::Tensor> = tokenized_input.
             iter().
